MarketingChatBot-main/requirements.txt
langchain
streamlit
streamlit_chat
pandas
openai
faiss-cpu
tiktoken
langchain-community
langchain-openai
python-dotenv